Stuck while toggling and shuffling rows, result in duplicate rows
