+++ app/src/main/java/com/example/canstudy/db/SwipeToDeleteCallback.kt
@@ -1,0 +1,28 @@
+package com.example.canstudy.db
+
+import androidx.recyclerview.widget.ItemTouchHelper
+import androidx.recyclerview.widget.RecyclerView
+
+/**
+ * A class that defines the swipe functionality for the item rows in the RecyclerView.
+ */
+
+abstract class SwipeToDeleteCallback: ItemTouchHelper.Callback() {
+
+    override fun getMovementFlags(
+        recyclerView: RecyclerView,
+        viewHolder: RecyclerView.ViewHolder
+    ): Int {
+        val swipeFlag = ItemTouchHelper.LEFT or ItemTouchHelper.RIGHT
+        return makeMovementFlags(0, swipeFlag)
+    }
+
+    // This function does nothing as we only want the swipe functionality.
+    override fun onMove(
+        recyclerView: RecyclerView,
+        viewHolder: RecyclerView.ViewHolder,
+        target: RecyclerView.ViewHolder
+    ): Boolean {
+        return false
+    }
+}
